oncall_env/server/requirements.txt
openenv-core[core]>=0.2.2
openai>=1.0.0
uvicorn==0.44.0
sse-starlette==3.3.4
starlette==1.0.0